Remove special build tag to test errors 1.13 functions. Go 1.13 or higher is now required
